requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
pandas>=2.2.0
plotly>=5.19.0
openpyxl>=3.1.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
slack-sdk>=3.27.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
